scripts/requirements.txt
PyGithub>=1.55.0
openai>=1.0.0
tiktoken>=0.4.0
google-generativeai>=0.3.0
